# fifocache: use fibonacci hashing