## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #21393

## What this PR does / why we need it:
Make hashing much faster.

benchstat:

```
goos: linux
goarch: amd64
pkg: github.com/matrixorigin/matrixone/pkg/fileservice/fifocache
cpu: AMD Ryzen 9 7900 12-Core Processor             
                 │      b1       │                 b2                  │
                 │    sec/op     │    sec/op     vs base               │
ShardCacheKey-24   21.1650n ± 1%   0.1966n ± 6%  -99.07% (p=0.002 n=6)

goos: linux
goarch: amd64
pkg: github.com/matrixorigin/matrixone/pkg/fileservice/fifocache
cpu: AMD Ryzen 9 7900 12-Core Processor             
                │     b1      │                 b2                 │
                │   sec/op    │   sec/op     vs base               │
DataCacheSet-24   260.4n ± 6%   193.7n ± 4%  -25.60% (p=0.002 n=6)
DataCacheGet-24   36.74n ± 1%   18.88n ± 2%  -48.61% (p=0.002 n=6)
geomean           97.82n        60.49n       -38.16%

```